Add Minimal Thread Device to build.
