+++ src/core/Makefile.am
@@ -28,14 +28,17 @@
 
 include $(abs_top_nlbuild_autotools_dir)/automake/pre.am
 
-lib_LIBRARIES                       = libopenthread.a
+lib_LIBRARIES                       = \
+    libopenthread-ftd.a               \
+    libopenthread-mtd.a               \
+    $(NULL)
 
-libopenthread_a_CPPFLAGS            = \
+CPPFLAGS_COMMON                     = \
     -I$(top_srcdir)/include           \
     $(OPENTHREAD_TARGET_DEFINES)      \
     $(NULL)
 
-libopenthread_a_SOURCES             = \
+SOURCES_COMMON                      = \
     openthread.cpp                    \
     coap/coap_base.cpp                \
     coap/coap_client.cpp              \
@@ -57,8 +60,6 @@ libopenthread_a_SOURCES             = \
     mac/mac_frame.cpp                 \
     mac/mac_whitelist.cpp             \
     mac/mac_blacklist.cpp             \
-    meshcop/joiner_router.cpp         \
-    meshcop/leader.cpp                \
     net/icmp6.cpp                     \
     net/ip6.cpp                       \
     net/ip6_address.cpp               \
@@ -67,7 +68,6 @@ libopenthread_a_SOURCES             = \
     net/ip6_routes.cpp                \
     net/netif.cpp                     \
     net/udp6.cpp                      \
-    thread/address_resolver.cpp       \
     thread/announce_begin_server.cpp  \
     thread/energy_scan_server.cpp     \
     thread/key_manager.cpp            \
@@ -78,10 +78,8 @@ libopenthread_a_SOURCES             = \
     thread/meshcop_tlvs.cpp           \
     thread/mesh_forwarder.cpp         \
     thread/mle.cpp                    \
-    thread/mle_router.cpp             \
     thread/mle_tlvs.cpp               \
     thread/network_data.cpp           \
-    thread/network_data_local.cpp     \
     thread/network_data_leader.cpp    \
     thread/panid_query_server.cpp     \
     thread/network_diag.cpp           \
@@ -92,7 +90,7 @@ libopenthread_a_SOURCES             = \
     $(NULL)
 
 if OPENTHREAD_ENABLE_COMMISSIONER
-libopenthread_a_SOURCES            += \
+SOURCES_COMMON                     += \
     coap/secure_coap_server.cpp       \
     meshcop/announce_begin_client.cpp \
     meshcop/commissioner.cpp          \
@@ -102,35 +100,57 @@ libopenthread_a_SOURCES            += \
 endif  # OPENTHREAD_ENABLE_COMMISSIONER
 
 if OPENTHREAD_ENABLE_JOINER
-libopenthread_a_SOURCES            += \
+SOURCES_COMMON                     += \
+    coap/secure_coap_client.cpp       \
     meshcop/joiner.cpp                \
     $(NULL)
 endif  # OPENTHREAD_ENABLE_JOINER
 
 if OPENTHREAD_ENABLE_DTLS
-libopenthread_a_SOURCES            += \
-    coap/secure_coap_client.cpp       \
+SOURCES_COMMON                     += \
     meshcop/dtls.cpp                  \
     $(NULL)
 endif  # OPENTHREAD_ENABLE_DTLS
 
 if OPENTHREAD_ENABLE_DHCP6_CLIENT
-libopenthread_a_SOURCES            += \
+SOURCES_COMMON                     += \
     net/dhcp6_client.cpp              \
     $(NULL)
 endif  # OPENTHREAD_ENABLE_DHCP6_CLIENT
 
 if OPENTHREAD_ENABLE_DHCP6_SERVER
-libopenthread_a_SOURCES            += \
+SOURCES_COMMON                     += \
     net/dhcp6_server.cpp              \
     $(NULL)
 endif  # OPENTHREAD_ENABLE_DHCP6_SERVER
 
 if OPENTHREAD_ENABLE_JAM_DETECTION
-libopenthread_a_SOURCES            += \
+SOURCES_COMMON                     += \
     utils/jam_detector.cpp            \
     $(NULL)
 endif  # OPENTHREAD_ENABLE_DHCP6_SERVER
+
+libopenthread_ftd_a_CPPFLAGS        = \
+    $(CPPFLAGS_COMMON)                \
+    $(NULL)
+
+libopenthread_ftd_a_SOURCES         = \
+    $(SOURCES_COMMON)                 \
+    meshcop/joiner_router.cpp         \
+    meshcop/leader.cpp                \
+    thread/address_resolver.cpp       \
+    thread/mle_router.cpp             \
+    thread/network_data_local.cpp     \
+    $(NULL)
+
+libopenthread_mtd_a_CPPFLAGS        = \
+    $(CPPFLAGS_COMMON)                \
+    -DOPENTHREAD_MTD                  \
+    $(NULL)
+
+libopenthread_mtd_a_SOURCES         = \
+    $(SOURCES_COMMON)                 \
+    $(NULL)
 
 noinst_HEADERS                      = \
     openthread-core-config.h          \
@@ -168,7 +188,11 @@ noinst_HEADERS                      = \
     meshcop/energy_scan_client.hpp    \
     meshcop/joiner.hpp                \
     meshcop/joiner_router.hpp         \
+    meshcop/joiner_router_ftd.hpp     \
+    meshcop/joiner_router_mtd.hpp     \
     meshcop/leader.hpp                \
+    meshcop/leader_ftd.hpp            \
+    meshcop/leader_mtd.hpp            \
     meshcop/panid_query_client.hpp    \
     net/icmp6.hpp                     \
     net/ip6.hpp                       \
@@ -185,6 +209,8 @@ noinst_HEADERS                      = \
     net/dhcp6_client.hpp              \
     net/dhcp6_server.hpp              \
     thread/address_resolver.hpp       \
+    thread/address_resolver_ftd.hpp   \
+    thread/address_resolver_mtd.hpp   \
     thread/announce_begin_server.hpp  \
     thread/energy_scan_server.hpp     \
     thread/key_manager.hpp            \
@@ -198,10 +224,14 @@ noinst_HEADERS                      = \
     thread/mle.hpp                    \
     thread/mle_constants.hpp          \
     thread/mle_router.hpp             \
+    thread/mle_router_ftd.hpp         \
+    thread/mle_router_mtd.hpp         \
     thread/mle_tlvs.hpp               \
     thread/network_data.hpp           \
     thread/network_data_leader.hpp    \
     thread/network_data_local.hpp     \
+    thread/network_data_local_ftd.hpp \
+    thread/network_data_local_mtd.hpp \
     thread/network_data_tlvs.hpp      \
     thread/panid_query_server.hpp     \
     thread/network_diag.hpp           \
