+++ src/core/meshcop/joiner_router.hpp
@@ -26,85 +26,8 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  */
 
-/**
- * @file
- *  This file includes definitions for the Joiner Router role.
- */
-
-#ifndef JOINER_ROUTER_HPP_
-#define JOINER_ROUTER_HPP_
-
-#include <openthread-types.h>
-
-#include <coap/coap_header.hpp>
-#include <coap/coap_client.hpp>
-#include <coap/coap_server.hpp>
-#include <common/message.hpp>
-#include <mac/mac_frame.hpp>
-#include <net/udp6.hpp>
-
-namespace Thread {
-
-class ThreadNetif;
-
-namespace MeshCoP {
-
-class JoinerRouter
-{
-public:
-    /**
-     * This constructor initializes the Joiner Router object.
-     *
-     * @param[in]  aThreadNetif  A reference to the Thread network interface.
-     *
-     */
-    JoinerRouter(ThreadNetif &aNetif);
-
-    /**
-     * This method returns the Joiner UDP Port.
-     *
-     * @returns The Joiner UDP Port number .
-     *
-     */
-    uint16_t GetJoinerUdpPort(void);
-
-    /**
-     * This method sets the Joiner UDP Port.
-     *
-     * @param[in]  The Joiner UDP Port number.
-     *
-     * @retval kThreadError_None    Successfully set the Joiner UDP Port.
-     *
-     */
-    ThreadError SetJoinerUdpPort(uint16_t aJoinerUdpPort);
-
-private:
-    static void HandleNetifStateChanged(uint32_t aFlags, void *aContext);
-    void HandleNetifStateChanged(uint32_t aFlags);
-
-    static void HandleUdpReceive(void *aContext, otMessage aMessage, const otMessageInfo *aMessageInfo);
-    void HandleUdpReceive(Message &aMessage, const Ip6::MessageInfo &aMessageInfo);
-
-    static void HandleRelayTransmit(void *aContext, otCoapHeader *aHeader, otMessage aMessage,
-                                    const otMessageInfo *aMessageInfo);
-    void HandleRelayTransmit(Coap::Header &aHeader, Message &aMessage, const Ip6::MessageInfo &aMessageInfo);
-
-    ThreadError SendJoinerEntrust(const Ip6::MessageInfo &aMessageInfo);
-
-    ThreadError GetBorderAgentRloc(uint16_t &aRloc);
-
-    Ip6::NetifCallback mNetifCallback;
-
-    Ip6::UdpSocket mSocket;
-    Coap::Resource mRelayTransmit;
-    Coap::Client &mCoapClient;
-    ThreadNetif &mNetif;
-
-    uint16_t mJoinerUdpPort;
-    bool mIsJoinerPortConfigured;
-};
-
-}  // namespace MeshCoP
-}  // namespace Thread
-
-#endif  // JOINER_ROUTER_HPP_
+#ifdef OPENTHREAD_MTD
+#include "joiner_router_mtd.hpp"
+#else
+#include "joiner_router_ftd.hpp"
+#endif
