+++ src/core/meshcop/leader.hpp
@@ -26,98 +26,8 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  */
 
-/**
- * @file
- *   This file includes definitions for a MeshCoP Leader.
- */
-
-#ifndef MESHCOP_LEADER_HPP_
-#define MESHCOP_LEADER_HPP_
-
-#include <coap/coap_client.hpp>
-#include <coap/coap_server.hpp>
-#include <common/timer.hpp>
-#include <net/udp6.hpp>
-#include <thread/meshcop_tlvs.hpp>
-#include <thread/mle.hpp>
-
-namespace Thread {
-namespace MeshCoP {
-
-OT_TOOL_PACKED_BEGIN
-class CommissioningData
-{
-public:
-    uint8_t GetLength(void) const {
-        return sizeof(Tlv) + mBorderAgentLocator.GetLength() +
-               sizeof(Tlv) + mCommissionerSessionId.GetLength() +
-               sizeof(Tlv) + mSteeringData.GetLength();
-    }
-
-    BorderAgentLocatorTlv mBorderAgentLocator;
-    CommissionerSessionIdTlv mCommissionerSessionId;
-    SteeringDataTlv mSteeringData;
-} OT_TOOL_PACKED_END;
-
-class Leader
-{
-public:
-    /**
-     * This constructor initializes the Leader object.
-     *
-     * @param[in]  aThreadNetif  A reference to the Thread network interface.
-     *
-     */
-    Leader(ThreadNetif &aThreadNetif);
-
-    /**
-     * This method sends a MGMT_DATASET_CHANGED message to commissioner.
-     *
-     * @param[in]  aAddress   The IPv6 address of destination.
-     *
-     * @retval kThreadError_None    Successfully send MGMT_DATASET_CHANGED message.
-     * @retval kThreadError_NoBufs  Insufficient buffers to generate a MGMT_DATASET_CHANGED message.
-     *
-     */
-    ThreadError SendDatasetChanged(const Ip6::Address &aAddress);
-
-private:
-    enum
-    {
-        kTimeoutLeaderPetition = 50, ///< TIMEOUT_LEAD_PET (seconds)
-    };
-
-    static void HandleTimer(void *aContext);
-    void HandleTimer(void);
-
-    static void HandlePetition(void *aContext, otCoapHeader *aHeader, otMessage aMessage,
-                               const otMessageInfo *aMessageInfo);
-    void HandlePetition(Coap::Header &aHeader, Message &aMessage, const Ip6::MessageInfo &aMessageInfo);
-    ThreadError SendPetitionResponse(const Coap::Header &aRequestHeader, const Ip6::MessageInfo &aMessageInfo,
-                                     StateTlv::State aState);
-
-    static void HandleKeepAlive(void *aContext, otCoapHeader *aHeader, otMessage aMessage,
-                                const otMessageInfo *aMessageInfo);
-    void HandleKeepAlive(Coap::Header &aHeader, Message &aMessage, const Ip6::MessageInfo &aMessageInfo);
-    ThreadError SendKeepAliveResponse(const Coap::Header &aRequestHeader, const Ip6::MessageInfo &aMessageInfo,
-                                      StateTlv::State aState);
-
-    static void HandleUdpReceive(void *aContext, otMessage aMessage, const otMessageInfo *aMessageInfo);
-
-    Coap::Resource mPetition;
-    Coap::Resource mKeepAlive;
-    Coap::Server &mCoapServer;
-    Coap::Client &mCoapClient;
-    NetworkData::Leader &mNetworkData;
-
-    Timer mTimer;
-
-    CommissionerIdTlv mCommissionerId;
-    uint16_t mSessionId;
-    ThreadNetif &mNetif;
-};
-
-}  // namespace MeshCoP
-}  // namespace Thread
-
-#endif  // MESHCOP_LEADER_HPP_
+#ifdef OPENTHREAD_MTD
+#include "leader_mtd.hpp"
+#else
+#include "leader_ftd.hpp"
+#endif
