+++ src/core/thread/network_data_local.hpp
@@ -26,129 +26,8 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  */
 
-/**
- * @file
- *   This file includes definitions for manipulating local Thread Network Data.
- */
-
-#ifndef NETWORK_DATA_LOCAL_HPP_
-#define NETWORK_DATA_LOCAL_HPP_
-
-#include <thread/network_data.hpp>
-
-namespace Thread {
-
-class ThreadNetif;
-
-/**
- * @addtogroup core-netdata-local
- *
- * @brief
- *   This module includes definitions for manipulating local Thread Network Data.
- *
- * @{
- */
-
-namespace NetworkData {
-
-/**
- * This class implements the Thread Network Data contributed by the local device.
- *
- */
-class Local: public NetworkData
-{
-public:
-    /**
-     * This constructor initializes the local Network Data.
-     *
-     * @param[in]  aNetif  A reference to the Thread network interface.
-     *
-     */
-    explicit Local(ThreadNetif &aNetif);
-
-    /**
-     * This method adds a Border Router entry to the Thread Network Data.
-     *
-     * @param[in]  aPrefix        A pointer to the prefix.
-     * @param[in]  aPrefixLength  The length of @p aPrefix in bytes.
-     * @param[in]  aPrf           The preference value.
-     * @param[in]  aFlags         The Border Router Flags value.
-     * @param[in]  aStable        The Stable value.
-     *
-     * @retval kThreadError_None    Successfully added the Border Router entry.
-     * @retval kThreadError_NoBufs  Insufficient space to add the Border Router entry.
-     *
-     */
-    ThreadError AddOnMeshPrefix(const uint8_t *aPrefix, uint8_t aPrefixLength, int8_t aPrf, uint8_t aFlags,
-                                bool aStable);
-
-    /**
-     * This method removes a Border Router entry from the Thread Network Data.
-     *
-     * @param[in]  aPrefix        A pointer to the prefix.
-     * @param[in]  aPrefixLength  The length of @p aPrefix in bytes.
-     *
-     * @retval kThreadError_None      Successfully removed the Border Router entry.
-     * @retval kThreadError_NotFound  Could not find the Border Router entry.
-     *
-     */
-    ThreadError RemoveOnMeshPrefix(const uint8_t *aPrefix, uint8_t aPrefixLength);
-
-    /**
-     * This method adds a Has Route entry to the Thread Network data.
-     *
-     * @param[in]  aPrefix        A pointer to the prefix.
-     * @param[in]  aPrefixLength  The length of @p aPrefix in bytes.
-     * @param[in]  aPrf           The preference value.
-     * @param[in]  aStable        The Stable value.
-     *
-     * @retval kThreadError_None    Successfully added the Has Route entry.
-     * @retval kThreadError_NoBufs  Insufficient space to add the Has Route entry.
-     *
-     */
-    ThreadError AddHasRoutePrefix(const uint8_t *aPrefix, uint8_t aPrefixLength, int8_t aPrf, bool aStable);
-
-    /**
-     * This method removes a Border Router entry from the Thread Network Data.
-     *
-     * @param[in]  aPrefix        A pointer to the prefix.
-     * @param[in]  aPrefixLength  The length of @p aPrefix in bytes.
-     *
-     * @retval kThreadError_None      Successfully removed the Border Router entry.
-     * @retval kThreadError_NotFound  Could not find the Border Router entry.
-     *
-     */
-    ThreadError RemoveHasRoutePrefix(const uint8_t *aPrefix, uint8_t aPrefixLength);
-
-    /**
-     * This method sends a Server Data Notification message to the Leader.
-     *
-     * @retval kThreadError_None    Successfully enqueued the notification message.
-     * @retval kThreadError_NoBufs  Insufficient message buffers to generate the notification message.
-     *
-     */
-    ThreadError SendServerDataNotification(void);
-
-private:
-    ThreadError UpdateRloc(void);
-    ThreadError UpdateRloc(PrefixTlv &aPrefix);
-    ThreadError UpdateRloc(HasRouteTlv &aHasRoute);
-    ThreadError UpdateRloc(BorderRouterTlv &aBorderRouter);
-
-    bool IsOnMeshPrefixConsistent(void);
-    bool IsExternalRouteConsistent(void);
-
-    uint16_t mOldRloc;
-
-    Leader &mLeader;
-};
-
-}  // namespace NetworkData
-
-/**
- * @}
- */
-
-}  // namespace Thread
-
-#endif  // NETWORK_DATA_LOCAL_HPP_
+#ifdef OPENTHREAD_MTD
+#include "network_data_local_mtd.hpp"
+#else
+#include "network_data_local_ftd.hpp"
+#endif
